requirements.txt
pandas>=2.3.3
openai
python-dotenv
apsw
sqlite-vec


